Add transformations
This update adds Transformations, a way to transform the ctx
value that is passed along with some small refactoring that was
needed to be used for a different project
