+++ package.json
@@ -13,6 +13,7 @@
     "start": "NODE_ENV=production node src/index.js"
   },
   "devDependencies": {
+    "concurrently": "^4.0.1",
     "dotenv": "^6.1.0",
     "eslint": "^5.7.0",
     "eslint-plugin-prettier": "^3.0.0",
