+++ src/configuration/index.js
@@ -1,3 +1,3 @@
 module.exports = {
-  routes: ['users', 'todos', 'lists', 'reminders', 'records', 'events']
+  routes: ['asks']
 }
