+++ src/queries/asks.js
@@ -1,0 +1,11 @@
+const router = require('../router.js')()
+
+const index = (ctx, next) => {
+  return next()
+}
+
+router.resource('asks', {
+  index
+})
+
+module.exports = router
