+++ src/router.js
@@ -1,0 +1,6 @@
+const router = require('koa-rest-router')
+
+module.exports = () =>
+  router({
+    prefix: process.env.REST_PREFIX || ''
+  })
